Commit Wed08/31/2022

+++ src/JTCWH/JT4Exercise1.java
@@ -10,21 +10,21 @@ public class JT4Exercise1 {
         String studentName = student.nextLine();
 
         System.out.print("The marks in Subject 1 out of 100 is: ");
-        int subject1 = student.nextInt();
+        float subject1 = student.nextFloat();
 
         System.out.print("The marks in Subject 2 out of 100 is: ");
-        int subject2 = student.nextInt();
+        float subject2 = student.nextFloat();
 
         System.out.print("The marks in Subject 3 out of 100 is: ");
-        int subject3 = student.nextInt();
+        float subject3 = student.nextFloat();
 
         System.out.print("The marks in Subject 4 out of 100 is: ");
-        int subject4 = student.nextInt();
+        float subject4 = student.nextFloat();
 
         System.out.print("The marks in Subject 5 out of 100 is: ");
-        int subject5 = student.nextInt();
+        float subject5 = student.nextFloat();
 
-        int percentage = ((subject1 + subject2 + subject3 + subject4 + subject5) * 100) / 500;
+        float percentage = ((subject1 + subject2 + subject3 + subject4 + subject5) * 100) / 500;
 
         System.out.println("The calculated percentage of " + studentName + " is " + percentage + "%.");
